backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pytest
httpx
sqlalchemy
alembic
psycopg[binary]
passlib[bcrypt]
bcrypt<5
python-jose[cryptography]
email-validator
